+++ ichnaea/tests/base.py
@@ -424,11 +424,13 @@ class DBTestCase(LogTestCase):
             conn.execute(ApiKey.__table__.delete())
 
             key1 = ApiKey.__table__.insert().values(
-                valid_key='test', shortname='test', allow_fallback=False,
+                valid_key='test', shortname='test',
+                allow_fallback=False, allow_locate=True,
                 log_locate=True, log_region=True, log_submit=True)
             conn.execute(key1)
             key2 = ApiKey.__table__.insert().values(
-                valid_key='export', shortname='export', allow_fallback=False,
+                valid_key='export', shortname='export',
+                allow_fallback=False, allow_locate=False,
                 log_locate=False, log_region=False, log_submit=False)
             conn.execute(key2)